requirements.txt
Django==3.0.3
requests==2.23.0
djangorestframework==3.11.0
